(Refacto) Move supportedTypes.h to data namespace (and folder).
