+++ gegelatilib/include/data/dataHandler.h
@@ -5,7 +5,7 @@
 #include <vector>
 #include <functional>
 
-#include "supportedTypes.h"
+#include "data/supportedTypes.h"
 
 namespace std {
 	/*
