+++ gegelatilib/include/data/supportedTypes.h
