+++ gegelatilib/include/instructions/instruction.h
@@ -6,7 +6,7 @@
 #include <functional>
 
 #include "parameter.h"
-#include "supportedTypes.h"
+#include "data/supportedTypes.h"
 
 namespace Instructions {
 	/**
